+++ src/main/java/net/geforcemods/securitycraft/entity/camera/CameraController.java
@@ -1,11 +1,25 @@
 package net.geforcemods.securitycraft.entity.camera;
 
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Deque;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
+
+import com.google.common.util.concurrent.AtomicDouble;
+import com.mojang.blaze3d.pipeline.RenderTarget;
+import com.mojang.blaze3d.pipeline.TextureTarget;
 
 import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.blockentities.FrameBlockEntity;
 import net.geforcemods.securitycraft.blockentities.SecurityCameraBlockEntity;
 import net.geforcemods.securitycraft.blocks.SecurityCameraBlock;
 import net.geforcemods.securitycraft.misc.KeyBindings;
@@ -22,12 +36,18 @@ import net.minecraft.client.CameraType;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
-import net.minecraft.client.multiplayer.ClientChunkCache;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.client.renderer.ShaderProgram;
+import net.minecraft.client.renderer.chunk.SectionRenderDispatcher.CompiledSection;
+import net.minecraft.client.renderer.chunk.SectionRenderDispatcher.RenderSection;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.GlobalPos;
 import net.minecraft.core.SectionPos;
 import net.minecraft.network.protocol.game.ServerboundMovePlayerPacket;
+import net.minecraft.server.level.ChunkTrackingView;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
@@ -44,7 +64,6 @@ import net.neoforged.neoforge.network.PacketDistributor;
 public class CameraController {
 	public static CameraType previousCameraType;
 	public static boolean resetOverlaysAfterDismount = false;
-	private static ClientChunkCache.Storage cameraStorage;
 	private static final ViewMovementKeyHandler[] MOVE_KEY_HANDLERS = Util.make(() -> {
 		Minecraft mc = Minecraft.getInstance();
 
@@ -62,6 +81,12 @@ public class CameraController {
 			return new ViewMovementKeyHandler[0];
 	});
 	private static int screenshotSoundCooldown = 0;
+	public static final Map<GlobalPos, Set<BlockPos>> FRAME_LINKS = new HashMap<>();
+	public static final Map<GlobalPos, CameraFeed> FRAME_CAMERA_FEEDS = new ConcurrentHashMap<>();
+	public static final Set<GlobalPos> FEED_FRUSTUM_UPDATE_REQUIRED = new HashSet<>();
+	public static GlobalPos currentlyCapturedCamera;
+	public static ShaderProgram cameraMonitorShader;
+	public static double lastFrameRendered = 0.0D;
 
 	private CameraController() {}
 
@@ -215,21 +240,109 @@ public class CameraController {
 		PacketDistributor.sendToServer(new SetDefaultCameraViewingDirection(cam.getId(), cam.getXRot(), cam.getYRot(), cam.getZoomAmount()));
 	}
 
-	public static ClientChunkCache.Storage getCameraStorage() {
-		return cameraStorage;
+	public static boolean isLinked(FrameBlockEntity be, GlobalPos cameraPos) {
+		return FRAME_LINKS.containsKey(cameraPos) && FRAME_LINKS.get(cameraPos).contains(be.getBlockPos());
 	}
 
-	public static void setCameraStorage(ClientChunkCache.Storage newStorage) {
-		cameraStorage = newStorage;
+	public static void addFrameLink(FrameBlockEntity be, GlobalPos cameraPos) {
+		Set<BlockPos> bes = FRAME_LINKS.computeIfAbsent(cameraPos, p -> new HashSet<>());
+
+		bes.add(be.getBlockPos());
+		FRAME_CAMERA_FEEDS.computeIfAbsent(cameraPos, CameraController::setUpCameraSections);
 	}
 
-	public static void setRenderPosition(Entity entity) {
-		if (entity instanceof SecurityCamera) {
-			SectionPos cameraPos = SectionPos.of(entity);
+	private static CameraFeed setUpCameraSections(GlobalPos cameraPos) {
+		int resolution = ConfigHandler.CLIENT.frameFeedResolution.get();
+		BlockPos pos = cameraPos.pos();
+		SectionPos cameraSectionPos = SectionPos.of(pos);
+		RenderSection startingSection = CameraViewAreaExtension.rawFetch(cameraSectionPos.x(), Mth.clamp(cameraSectionPos.y(), CameraViewAreaExtension.minSectionY(), CameraViewAreaExtension.maxSectionY() - 1), cameraSectionPos.z(), true);
+		CameraFeed cameraFeed = new CameraFeed(new TextureTarget(resolution, resolution, true), new AtomicDouble(), new ArrayList<>(), new HashSet<>(), new ArrayList<>(), new ArrayList<>());
 
-			cameraStorage.viewCenterX = cameraPos.x();
-			cameraStorage.viewCenterZ = cameraPos.z();
+		cameraFeed.compilingSectionsQueue.add(startingSection);
+		cameraFeed.sectionsInRange.add(startingSection);
+		cameraFeed.sectionsInRangePositions.add(startingSection.getOrigin().asLong());
+		CameraController.discoverVisibleSections(cameraPos, getFrameFeedViewDistance(null), cameraFeed);
+		return cameraFeed;
+	}
+
+	public static void removeFrameLink(FrameBlockEntity be, GlobalPos cameraPos) {
+		if (FRAME_LINKS.containsKey(cameraPos)) {
+			Set<BlockPos> linkedFrames = FRAME_LINKS.get(cameraPos);
+
+			linkedFrames.remove(be.getBlockPos());
+
+			if (linkedFrames.isEmpty())
+				removeAllFrameLinks(cameraPos);
 		}
+	}
+
+	public static void removeAllFrameLinks(GlobalPos cameraPos) {
+		if (FRAME_LINKS.containsKey(cameraPos)) {
+			FRAME_LINKS.remove(cameraPos);
+			FRAME_CAMERA_FEEDS.remove(cameraPos);
+		}
+	}
+
+	public static RenderTarget getViewForFrame(GlobalPos cameraPos) {
+		return FRAME_CAMERA_FEEDS.containsKey(cameraPos) ? FRAME_CAMERA_FEEDS.get(cameraPos).renderTarget : null;
+	}
+
+	public static void discoverVisibleSections(GlobalPos cameraPos, int viewDistance, CameraFeed feed) {
+		SectionPos cameraSectionPos = SectionPos.of(cameraPos.pos());
+		List<RenderSection> visibleSections = feed.sectionsInRange;
+		List<RenderSection> sectionQueue = feed.compilingSectionsQueue;
+		Set<Long> visibleSectionPositions = feed.sectionsInRangePositions;
+		Deque<RenderSection> queueToCheck = new ArrayDeque<>(sectionQueue);
+
+		sectionQueue.clear();
+
+		while (!queueToCheck.isEmpty()) {
+			RenderSection currentSection = queueToCheck.poll();
+			BlockPos origin = currentSection.getOrigin();
+			CompiledSection currentCompiledSection = currentSection.getCompiled();
+
+			if (currentCompiledSection == CompiledSection.UNCOMPILED) {
+				sectionQueue.add(currentSection);
+				continue;
+			}
+
+			//Once a section in the queue is compiled, it knows which neighbours it can and cannot see. This information is used to more cleverly determine which chunks the player can actually see
+			for (Direction dir : Direction.values()) {
+				int cx = SectionPos.blockToSectionCoord(origin.getX()) + dir.getStepX();
+				int cy = SectionPos.blockToSectionCoord(origin.getY()) + dir.getStepY();
+				int cz = SectionPos.blockToSectionCoord(origin.getZ()) + dir.getStepZ();
+
+				if (ChunkTrackingView.isInViewDistance(cameraSectionPos.x(), cameraSectionPos.z(), viewDistance, cx, cz)) {
+					RenderSection neighbourSection = CameraViewAreaExtension.rawFetch(cx, cy, cz, true);
+
+					if (neighbourSection != null) {
+						long neighbourPosAsLong = neighbourSection.getOrigin().asLong();
+
+						if (!visibleSectionPositions.contains(neighbourPosAsLong) && canSeeNeighborFace(currentCompiledSection, dir)) {
+							visibleSections.add(neighbourSection); //Yet uncompiled render sections are added to the sections-in-range list, so Minecraft will schedule to compile them
+							visibleSectionPositions.add(neighbourSection.getOrigin().asLong());
+							sectionQueue.add(neighbourSection);
+							FEED_FRUSTUM_UPDATE_REQUIRED.add(cameraPos);
+						}
+					}
+				}
+			}
+		}
+	}
+
+	private static boolean canSeeNeighborFace(CompiledSection currentCompiledSection, Direction dir) {
+		for (int j = 0; j < Direction.values().length; j++) {
+			if (currentCompiledSection.facesCanSeeEachother(Direction.values()[j].getOpposite(), dir))
+				return true;
+		}
+
+		return false;
+	}
+
+	public static int getFrameFeedViewDistance(FrameBlockEntity be) {
+		int frameSpecificRenderDistance = be == null ? 32 : be.getChunkLoadingDistanceOption();
+
+		return Math.min(frameSpecificRenderDistance, Math.min(ConfigHandler.CLIENT.frameFeedRenderDistance.get(), Math.min(ConfigHandler.SERVER.frameFeedViewDistance.get(), Minecraft.getInstance().options.getEffectiveRenderDistance())));
 	}
 
 	public static float getMovementSpeed(SecurityCamera cam) {
@@ -265,4 +378,6 @@ public class CameraController {
 			}
 		}
 	}
+
+	public record CameraFeed(RenderTarget renderTarget, AtomicDouble lastActiveTime, List<RenderSection> sectionsInRange, Set<Long> sectionsInRangePositions, List<RenderSection> visibleSections, List<RenderSection> compilingSectionsQueue) {}
 }